доделал до 10 шага
