+++ src/components/point.js
@@ -1,7 +1,5 @@
 // отрисовыает точки и даты маршрута
-import {
-  createElement
-} from '../utils.js';
+import AbstractComponent from "../components/abstract-component.js";
 
 const getPathPointTown = (itemArray) => {
   let totalPath = ``;
@@ -27,24 +25,15 @@ const getSitePathTemplate = (listEvent) => {
   );
 };
 
-export default class SitePathTemplate {
+export default class SitePathTemplate extends AbstractComponent {
   constructor(point) {
-    this._point = point;
+    super();
 
-    this._element = null;
+    this._point = point;
   }
 
   getTemplate() {
     return getSitePathTemplate(this._point);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
-  }
 }
